rename wasm zcap module
