+++ sdk-wasm/src/authorization.rs
